Created script for voronoi creationg in Qgis.
Turns a bunch of clicking into selecting the paths of  2 inputs and a path of one output folder. Have not tried running this outside Qgis probably does not work.
For some ungodly reason python input modules are broken inside so i used a Qgis module.
I have not cleaned up the code much, but it should hopefully work with whats inside the folder. Just copy the path and enter it 3 times.
More info in the readme